// src/app/features/stock-page/enums/stock-options.enum.ts
export enum StockFabricStatus {
     ALL = '',
     RETURNED = 'returned',
     RUNOUT = 'runout',
     ORDERED_MAT = 'ordered_mat',
}

export enum StockOptionTitles {
     ALL = 'Stan magazynu',
     RETURNED = 'Zwroty',
     RUNOUT = 'Artykuły wyczepane',
     ORDERED_MAT = 'Zamówione materiały',
}

export enum STOCK_ASSORTS_ID {
     FABRIC = 70,
     SHOES = 60,
     ADDS = 50,
     POCKET_SQUARE = 51,
     BOW_TIE = 53,
     TIE = 52,
}